# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: 연금술사 Prototype API
  version: 0.1.0
paths:
...
  /api/v1/projects/{project_id}/documents:
    get:
      summary: Documents
      operationId: documents_api_v1_projects__project_id__documents_get
      security:
      - HTTPBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      summary: Upload
      operationId: upload_api_v1_projects__project_id__documents_post
      security:
      - HTTPBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_api_v1_projects__project_id__documents_post'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/documents/{document_id}/versions:
    post:
      summary: Upload Version
      operationId: upload_version_api_v1_documents__document_id__versions_post
      security:
      - HTTPBearer: []
      parameters:
      - name: document_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Document Id
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_version_api_v1_documents__document_id__versions_post'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      summary: Document Versions
      operationId: document_versions_api_v1_documents__document_id__versions_get
      security:
      - HTTPBearer: []
      parameters:
      - name: document_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Document Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/documents/{document_id}:
    get:
      summary: Document Detail
      operationId: document_detail_api_v1_documents__document_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: document_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Document Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      summary: Delete Document
      operationId: delete_document_api_v1_documents__document_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: document_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Document Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/documents/{document_id}/download:
    get:
      summary: Document Download
      operationId: document_download_api_v1_documents__document_id__download_get
      security:
      - HTTPBearer: []
      parameters:
      - name: document_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Document Id
      - name: version_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Version Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/documents/{document_id}/analyze:
    post:
      summary: Analyze
      operationId: analyze_api_v1_documents__document_id__analyze_post
      security:
      - HTTPBearer: []
      parameters:
      - name: document_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Document Id
      - name: idempotency-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Idempotency-Key
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/documents/{document_id}/slide-plans:
    post:
      summary: New Plan
      operationId: new_plan_api_v1_documents__document_id__slide_plans_post
      security:
      - HTTPBearer: []
      parameters:
      - name: document_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Document Id
      - name: idempotency-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Idempotency-Key
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      summary: Plans
      operationId: plans_api_v1_documents__document_id__slide_plans_get
      security:
      - HTTPBearer: []
      parameters:
      - name: document_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Document Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/documents/{document_id}/parse-result:
    get:
      summary: Parse Result
      operationId: parse_result_api_v1_documents__document_id__parse_result_get
      security:
      - HTTPBearer: []
      parameters:
      - name: document_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Document Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/documents/{document_id}/chunks:
    get:
      summary: Chunks
      operationId: chunks_api_v1_documents__document_id__chunks_get
      security:
      - HTTPBearer: []
      parameters:
      - name: document_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Document Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/slide-plans/{plan_id}:
    get:
      summary: Plan Detail
      operationId: plan_detail_api_v1_slide_plans__plan_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: plan_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Plan Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      summary: Plan Edit
      operationId: plan_edit_api_v1_slide_plans__plan_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: plan_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Plan Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SlidePlanContract'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/slide-plans/{plan_id}/validate:
    post:
      summary: Plan Validate
      operationId: plan_validate_api_v1_slide_plans__plan_id__validate_post
      security:
      - HTTPBearer: []
      parameters:
      - name: plan_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Plan Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/slide-plans/{plan_id}/approve:
    post:
      summary: Plan Approve
      operationId: plan_approve_api_v1_slide_plans__plan_id__approve_post
      security:
      - HTTPBearer: []
      parameters:
      - name: plan_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Plan Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewInput'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/slide-plans/{plan_id}/generate-ppt:
    post:
      summary: Generate Ppt
      operationId: generate_ppt_api_v1_slide_plans__plan_id__generate_ppt_post
      security:
      - HTTPBearer: []
      parameters:
      - name: plan_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Plan Id
      - name: idempotency-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Idempotency-Key
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/slide-plans/{plan_id}/generate-video:
    post:
      summary: Generate Video
      operationId: generate_video_api_v1_slide_plans__plan_id__generate_video_post
      security:
      - HTTPBearer: []
      parameters:
      - name: plan_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Plan Id
      - name: idempotency-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Idempotency-Key
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/slide-plans/{plan_id}/generate-images:
    post:
      summary: Generate Images
      operationId: generate_images_api_v1_slide_plans__plan_id__generate_images_post
      security:
      - HTTPBearer: []
      parameters:
      - name: plan_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Plan Id
      - name: idempotency-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Idempotency-Key
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/projects/{project_id}/jobs:
    get:
      summary: Jobs
      operationId: jobs_api_v1_projects__project_id__jobs_get
      security:
      - HTTPBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/jobs/{job_id}:
    get:
      summary: Job Detail
      operationId: job_detail_api_v1_jobs__job_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/jobs/{job_id}/cancel:
    post:
      summary: Job Cancel
      operationId: job_cancel_api_v1_jobs__job_id__cancel_post
      security:
      - HTTPBearer: []
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/jobs/{job_id}/retry:
    post:
      summary: Job Retry
      operationId: job_retry_api_v1_jobs__job_id__retry_post
      security:
      - HTTPBearer: []
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Job Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RetryInput'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/queues:
    get:
      summary: Queues
      operationId: queues_api_v1_admin_queues_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
  /api/v1/admin/templates:
    post:
      summary: Template Upload
      operationId: template_upload_api_v1_admin_templates_post
      security:
      - HTTPBearer: []
      parameters:
      - name: project_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_template_upload_api_v1_admin_templates_post'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    Body_template_upload_api_v1_admin_templates_post:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
      type: object
      required:
      - file
      title: Body_template_upload_api_v1_admin_templates_post
    Body_upload_api_v1_projects__project_id__documents_post:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
      type: object
      required:
      - file
      title: Body_upload_api_v1_projects__project_id__documents_post
    Body_upload_version_api_v1_documents__document_id__versions_post:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
      type: object
      required:
      - file
      title: Body_upload_version_api_v1_documents__document_id__versions_post
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    RetryInput:
      properties:
        resume_from_step:
          anyOf:
          - type: string
          - type: 'null'
          title: Resume From Step
      additionalProperties: false
      type: object
      title: RetryInput
    ReviewInput:
      properties:
        decision:
          type: string
          enum:
          - approve
          - reject
          - request-change
          title: Decision
        comment:
          type: string
          maxLength: 2000
          title: Comment
          default: ''
        version:
          type: integer
          minimum: 1.0
          title: Version
      additionalProperties: false
      type: object
      required:
      - decision
      - version
      title: ReviewInput
...
    SlidePlanContract:
      properties:
        schema_version:
          type: string
          const: '1.0'
          title: Schema Version
          default: '1.0'
        plan_id:
          type: string
          format: uuid
          title: Plan Id
        document_versions:
          items:
            type: string
            format: uuid
          type: array
          minItems: 1
          title: Document Versions
        title:
          type: string
          maxLength: 120
          minLength: 1
          title: Title
        audience:
          type: string
          const: internal
          title: Audience
          default: internal
        slides:
          items:
            $ref: '#/components/schemas/PlannedSlide'
          type: array
          maxItems: 40
          minItems: 1
          title: Slides
        mock:
          type: boolean
          title: Mock
        provenance:
          additionalProperties: true
          type: object
          title: Provenance
      additionalProperties: false
      type: object
      required:
      - document_versions
      - title
      - slides
      - mock
      title: SlidePlanContract
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer